// src/app/business-logic/model/tasks/tasksGetAllExResponse.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.7
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Task } from './task';


export interface TasksGetAllExResponse {
    /**
     * List of tasks
     */
    tasks?: Array<Task>;
}