// src/app/business-logic/model/tasks/tasksUpdateResponse.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.9
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface TasksUpdateResponse {
    /**
     * Number of tasks updated (0 or 1)
     */
    updated?: number;
    /**
     * Updated fields names and values
     */
    fields?: object;
}