// src/app/business-logic/model/projects/projectsGetTaskParentsResponse.ts
/**
 * projects
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.12
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface ProjectsGetTaskParentsResponse {
    /**
     * The list of unique task parents sorted by their names
     */
    parents?: Array<object>;
}